# swagger/swagger.yaml
swagger: '2.0'
info:
  title: METAX REST API
  version: v1
host: metax.example.org
schemes:
  - https
basePath: /rest/v1/
produces:
  - application/json
paths:
...
  /files/{PID}/XML:
    get:
      summary: get XML metadata from file
      produces:
        - application/xml
        - application/json
      parameters:
        - name: PID
          in: path
          description: Persistent ID of file OR the internal pk
          required: true
          type: string
        - name: namespace
          in: query
          description: Namespace of the schema. If omitted, returns a list of xml metadata namespaces associated with the file.
          type: string
      responses:
        '200':
          description: single xml metadata (application/xml), or a list of namespaces (application/json)
        '404':
          description: file not found
      tags:
        - File API
    post:
      summary: create new XML metadata
      consumes:
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - name: PID
          in: path
          description: persistent ID of file OR the internal pk
          required: true
          type: string
        - name: namespace
          in: query
          description: namespace of the schema. If a xml metadata already existed with given namespace, an error is returned
          required: true
          type: string
        - name: body
          in: body
          description: The XML to create.
          required: true
      responses:
        '201':
          description: return created XML
        '400':
          description: parameters contained errors, response includes details (application/json)
      tags:
        - File API
    put:
      summary: update XML metadata
      consumes:
        - application/xml
      parameters:
        - name: PID
          in: path
          description: persistent ID of file OR the internal pk
          required: true
          type: string
        - name: namespace
          in: query
          description: namespace of the schema
          required: true
          type: string
        - name: body
          in: body
          description: "The XML to update"
          required: true
      responses:
        '204':
          description: XML saved
        '400':
          description: parameters contained errors, response includes details
        '404':
          description: file or namespace does not exist
      tags:
        - File API
    delete:
      summary: delete XML metadata
      parameters:
        - name: PID
          in: path
          description: persistent ID of file OR the internal pk
          required: true
          type: string
        - name: namespace
          in: query
          description: namespace of the schema
          required: true
          type: string
      responses:
        '204':
          description: XML is deleted
        '400':
          description: parameters contained errors, response includes details
        '404':
          description: file or namespace does not exist
      tags:
        - File API
  /files/datasets:
    post:
      summary: get datasets where files belong to
      description: |
        Note&#58; The method is invoked using POST, because there are limits to length of query
        parameters in GET. Also, some clients forcibly shove parameters in body in GET
        requests to query parameters, so using POST instead is more guaranteed to work.
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: a list of file id's (integers) or identifiers (strings).
          required: true
          type: string
      responses:
        '200':
          description: a list of urn_identifiers. if the files were not found to belong to any datasets, an empty list is returned
        '404':
          description: files given in the list not found
      tags:
        - File API
...
  /datasets:
    # some of the parameters and returned fields are for TPAS usage only
    get:
      summary: "list datasets"
      parameters:
        - name: latest
          in: query
          description: only return latest versions
          required: false
          type: boolean
        - name: owner_id
          in: query
          description: id of the person who owns the record in metax
          required: false
          type: string
        - name: user_created
          in: query
          description: id of the person who created the record in metax
          required: false
          type: string
        - name: curator
          in: query
          description: curator identifier (field research_dataset-> curator-> identifier)
          required: false
          type: string
        - name: state
          in: query
          description: TPAS state (field preservation_state). multiple states using OR-logic are queriable in the same request, e.g. state=5,6. see valid values from http://iow.csc.fi/model/mrd/CatalogRecord/ field preservation_state
          required: false
          type: string
        - name: offset
          in: query
          description: offset for paging
          required: false
          type: integer
        - name: limit
          in: query
          description: limit for paging
          required: false
          type: integer
          default: 10
      responses:
        "200":
          description: successful operation, returns a list of datasets containing full dataset objects including their data catalog and contract information. when using query parameters, search result can be an empty list.
          schema:
            $ref: '#/definitions/CatalogRecord'
      tags:
        - Dataset API
    post:
      summary: create new dataset metadata
      description: |
        # research_dataset identifiers
        The field <b>urn_identifier</b> is always generated for the user. <b>preferred_identifier</b> can be provided by the user, and if omitted, the value of <b>urn_identifier</b> is copied to it.
        # research_dataset read-only fields
        - urn_identifier
        - total_byte_size (generated from files in the db according to files listed in the given dataset)
      consumes:
        - application/json
      parameters:
      - in: "body"
        name: "body"
        description: "Object or a list of objects to create."
        required: true
        schema:
          $ref: '#/definitions/CatalogRecord'
      responses:
        '201':
          description: Returns the created object, or if a list was given, a list of objects and errors.
        '400':
          description: parameters contained errors, response includes details.
      tags:
        - Dataset API
    put:
      summary: bulk update
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "A list of objects to update."
          required: true
      responses:
        '200':
          description: Sucessful operation. Return values include a list of errors, if any.
        '400':
          description: All updates failed. A list of errors is returned.
      tags:
        - Dataset API
    patch:
      summary: bulk update partial
      description: |
        The payload must include a field that can be used to identify the resource being updated. Acceptable identifier fields are: <b>id, research_dataset->urn_identifier</b>
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "A list of (partial) objects to update."
          required: true
      responses:
        '200':
          description: Some or all objects were updated. Return values contain list of full updated objects, and may include a list of errors.
        '400':
          description: All updates failed. A list of errors is returned.
      tags:
        - Dataset API
...
definitions:
  CatalogRecord:
    $ref: 'https://raw.githubusercontent.com/CSCfi/metax-api/test/src/metax_api/api/rest/base/api_schemas/catalogrecord.json#/definitions/CatalogRecord'
...